--- team-project/src/components/LandingPage.js
@@ -1,44 +1,0 @@
-import React from "react";
-import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
-import { faMagnifyingGlass } from "@fortawesome/free-solid-svg-icons";
-
-const LandingPage = () => {
-  return (
-    <>
-      <section className="hero">
-        <div className="all-hero">
-          <div className="hero-content">
-            <p>Free high resolution images</p>
-            <h1>
-              Download astonishing images & photos for you or your business
-            </h1>
-          </div>
-          <div className="hero-input">
-            <input type="text" placeholder="Search for free photos" />
-            <FontAwesomeIcon icon={faMagnifyingGlass} className="search-icon" />
-          </div>
-
-          <div className="popular">
-            <p>Popular categories: </p>
-            <ul className="pop-categories">
-              <li>
-                <a href="/">Nature</a>
-              </li>
-              <li>
-                <a href="/">Wallpaper</a>
-              </li>
-              <li>
-                <a href="/">Cities</a>
-              </li>
-              <li>
-                <a href="/">Architecture</a>
-              </li>
-            </ul>
-          </div>
-        </div>
-      </section>
-    </>
-  );
-};
-
-export default LandingPage;
