+++ team-project/src/components/NavBar.css
@@ -7,7 +7,6 @@
 
 body {
   width: 100%;
-  background-color: #1f1f1f;
 }
 
 /* Navigation bar */
